Moved `IS7Service.cs` for structure consistency.
